pyproject.toml
[tool.poetry]
name = "pile-bot"
version = "0.1.0"
description = ""
authors = ["Artyom Grebenyuk <a@grebeniuk.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.9.0"
python-telegram-bot = "^13.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
